fix installing go

+++ bash/setup_golang.sh
@@ -1,7 +1,23 @@
 #!/bin/bash
 
+source functions.sh
+
 VERSION=1.15.2
 FILE=go$VERSION.linux-amd64.tar.gz
-wget https://golang.org/dl/$FILE
-sudo tar -C /usr/local -xzf $FILE
-rm $FILE
+
+if ! command -v go; then
+	wget https://golang.org/dl/$FILE
+	sudo tar -C /usr/local -xzf $FILE
+	rm $FILE
+else
+	GO_LOC=`which go`
+	echo "'go' already found at $GO_LOC"
+fi
+
+link_dep go.sh
+source $HOME/.bashrc
+
+# https://github.com/moovweb/gvm
+if [ ! -d "$HOME/.gvm" ]; then
+	bash < <(curl -s -S -L https://raw.githubusercontent.com/moovweb/gvm/master/binscripts/gvm-installer)
+fi
